Remove/fix existing tests and refactor helpers to new measureUtils.ts
